+++ src/five/pt/tests/test_engine.py
@@ -32,6 +32,14 @@ class TestPatches(Sandboxed, ZopeTestCase):
         template = template.__of__(self.folder)
         self.assertTrue('world' in template())
 
+    def test_pagetemplatefile_processing_instruction_skipped(self):
+        from Products.PageTemplates.PageTemplateFile import PageTemplateFile
+
+        # test rendering engine
+        template = PageTemplateFile(os.path.join(path, "pi.pt"))
+        template = template.__of__(self.folder)
+        self.assertIn('world', template())
+
     def test_zopepagetemplate(self):
         from Products.PageTemplates.ZopePageTemplate import manage_addPageTemplate
         template = manage_addPageTemplate(self.folder, 'test')
@@ -51,6 +59,18 @@ class TestPatches(Sandboxed, ZopeTestCase):
         # test commit
         import transaction
         transaction.commit()
+
+    def test_zopepagetemplate_processing_instruction_skipped(self):
+        from Products.PageTemplates.ZopePageTemplate import manage_addPageTemplate
+        template = manage_addPageTemplate(self.folder, 'test')
+
+        # aq-wrap before we proceed
+        template = template.__of__(self.folder)
+
+        # test rendering engine
+        template.write(open(os.path.join(path, "pi.pt")).read())
+        self.assertIn('world', template())
+
 
 def test_suite():
     import sys